Release 1.0.0: Fist project done
It works, is probably not perfect, but im running out of time, and i
consider what i did more than good enough
Is a good practice to initialize the db on docker build?
Well, i dont think so, but then again, you asked me something weird
Was i supposed to do 2 different dockerfiles to etl and queries?
Im not really sure, and too late to be asking.

I think the instructions are fool proff, just copy paste kinda stuff
but still, there's the fear, i didnt asked anybody to test it for me
and i have a mac, not sure if everything will work on linux.
I didnt new anything about docker, so it was nice to learn
The classes were slow till we started the aws thing, now im very behind
you're still reading? dont think so.
Have a nice day, enjoy watching my code, or dont.
If you want to comment anyting be free, the PR is open